backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
psycopg2-binary>=2.9,<3.0
django-cors-headers>=4.3,<5.0
pillow>=10.0,<11.0
pytesseract>=0.3,<1.0
python-decouple>=3.8,<4.0
requests>=2.31,<3.0
whitenoise>=6.5,<7.0
gunicorn>=21.2,<23.0
